Create simple help service and directive
This commit adds 2 components:

1) A Help Service (gaHelpRegistry). This service can be used
to get a html snippet corresponding to a given help id. Right
now, this html snippets are hard coded, but in the future, they
will come either from some template, url or service. It's a
requirement that the help texts will be used outside of the
application. This needs to be defined first.

2) A Help Directive (gaHelp). This directive can be used to
integrate a help button (currently it's a text) anywhere in
the page and associate it with a given help-id (configured
with ga-help attribute). When the button is clicked, the
help is displayed.

Displaying of help is not nice now. We have 2 options as
of now.

  - Display in modal windows (default)
  - Display in popover

By specifying `help=popover` in the URL parameter, you can
switch between those two options. We are using pure bootstrap
for this. It's likely that we need to find a better solution
to display the help.

For demo purposes, I have integrated this print directive in
two locations:

  - In the bottom bar (next to the coordinate system selection
    combobox), thus placed directly in the index.html
  - In the print section of the accordeion (next to the Layout
    text), thus placed inside a partial

It's far from done, I'm opening it up for review and discussion
and to see how it interacts with existing elements.